# VAULT-31907: Entity loading speedup