--- a/vault/identity_store_util.go
+++ b/vault/identity_store_util.go
@@ -491,6 +491,7 @@ LOOP:
 
 				ns, err := i.namespacer.NamespaceByID(ctx, entity.NamespaceID)
 				if err != nil {
+					tx.Abort()
 					return err
 				}
 				if ns == nil {
@@ -501,6 +502,7 @@ LOOP:
 						i.logger.Warn("deleting entity and its any existing aliases", "name", entity.Name, "namespace_id", entity.NamespaceID)
 						err = i.entityPacker.DeleteItem(ctx, entity.ID)
 						if err != nil {
+							tx.Abort()
 							return err
 						}
 					}
@@ -511,9 +513,11 @@ LOOP:
 				// Ensure that there are no entities with duplicate names
 				entityByName, err := i.MemDBEntityByName(nsCtx, entity.Name, false)
 				if err != nil {
+					tx.Abort()
 					return nil
 				}
 				if err := i.conflictResolver.ResolveEntities(ctx, entityByName, entity); err != nil && !i.disableLowerCasedNames {
+					tx.Abort()
 					return err
 				}
 
@@ -531,6 +535,7 @@ LOOP:
 
 				err = i.loadLocalAliasesForEntity(ctx, entity, localAliasBuckets)
 				if err != nil {
+					tx.Abort()
 					return fmt.Errorf("failed to load local aliases from storage: %v", err)
 				}
 
@@ -543,6 +548,7 @@ LOOP:
 				// Only update MemDB and don't hit the storage again
 				err = i.upsertEntityInTxn(nsCtx, tx, entity, nil, false)
 				if err != nil {
+					tx.Abort()
 					return fmt.Errorf("failed to update entity in MemDB: %w", err)
 				}
 				upsertedItems += toBeUpserted